CrudApiRepo (#818)
removed some of the Traits for Mango, got rid of statics on the GormEntity for queryList and others. these used for testing mosty and now need to use Domain.repo.method for some
interface ApiCrudRepo replaces old traits now, uses default methods for query and others
EntityResponder is gone, CrudApi does it.
CrudApiController is the new base controller, it talks to a CrudApi not the repo
RestRepoApiController just extends CrudApiController so we dont need to mass change in domain9 yet.
CrudApi wraps the repo now and is a generic way to opperate on repo with main logic centralized there but not polluting repo and not depending on http stuff fro controller.
CrudApi can be setup as bean for a domain for when we need to do something special, SyncJobCrudApi is good example. Should extend from DefaultCrudApi which is the impl for the CrudApi interface